+++ .gitignore
@@ -1,2 +1,5 @@
 .vscode/*
 venv/*
+/__pycache__
+/node_modules
+/package-lock.json
